fix pylint R0401 error
